rice/requirements.txt
asgiref==3.3.4
cycler==0.10.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
numpy==1.21.0
pandas==1.2.5
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
